+++ src/Request/People.php
@@ -772,7 +772,7 @@ class People extends RequestCollection
         if ($maxId !== null) {
             $data['max_id'] = $maxId;
         }
-        $response = $this->ig->internal->sendGraph('284797047912213050552583965568', $data, 'FollowersList', 'xdt_api__v1__friendships__followers', 'false', 'pando', true);
+        $response = $this->ig->internal->sendGraph('284797047912213050552583965568', $data, 'FollowersList', 'xdt_api__v1__friendships__followers', 'false', 'pando', true, true);
         $arr = $response->asArray();
         if (isset($arr['data'])) {
             $data = $arr['data'];
@@ -821,7 +821,7 @@ class People extends RequestCollection
         if ($maxId !== null) {
             $data['max_id'] = $maxId;
         }
-        $response = $this->ig->internal->sendGraph('161046392815797863596808012372', $data, 'FollowingList', 'xdt_api__v1__friendships__following', 'false', 'pando', true);
+        $response = $this->ig->internal->sendGraph('161046392815797863596808012372', $data, 'FollowingList', 'xdt_api__v1__friendships__following', 'false', 'pando', true, true);
         $arr = $response->asArray();
         if (isset($arr['data'])) {
             $data = $arr['data'];